Moving BoxableTokenizer to tantivy.
